# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a sample server celler server.",
        "title": "go-admin-api",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8000",
    "basePath": "/v1",
    "paths": {
...
        "/dictType/create": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "数据字典类型"
                ],
                "summary": "创建字典类型",
                "parameters": [
                    {
                        "description": "创建字典类型",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.CreateDictTypeRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "desc",
                        "schema": {
                            "$ref": "#/definitions/result.HttpResult"
                        }
                    }
                }
            }
        },
        "/dictType/delete/{id}": {
            "delete": {
                "tags": [
                    "数据字典类型"
                ],
                "summary": "删除字典类型",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "字典类型id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "desc",
                        "schema": {
                            "$ref": "#/definitions/result.HttpResult"
                        }
                    }
                }
            }
        },
        "/dictType/info/{id}": {
            "get": {
                "tags": [
                    "数据字典类型"
                ],
                "summary": "字典类型详情",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "字典类型id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "desc",
                        "schema": {
                            "$ref": "#/definitions/result.HttpResult"
                        }
                    }
                }
            }
        },
        "/dictType/page": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "数据字典类型"
                ],
                "summary": "字典类型列表",
                "parameters": [
                    {
                        "type": "integer",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "name": "size",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "desc",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/result.HttpResult"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "allOf": [
                                                {
                                                    "$ref": "#/definitions/response.PageResponse"
                                                },
                                                {
                                                    "type": "object",
                                                    "properties": {
                                                        "list": {
                                                            "type": "array",
                                                            "items": {
                                                                "$ref": "#/definitions/models.SysDictType"
                                                            }
                                                        }
                                                    }
                                                }
                                            ]
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/dictType/update": {
            "patch": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "数据字典类型"
                ],
                "summary": "更新字典类型",
                "parameters": [
                    {
                        "description": "更新字典类型",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.UpdateDictTypeRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "desc",
                        "schema": {
                            "$ref": "#/definitions/result.HttpResult"
                        }
                    }
                }
            }
        },
...
        "/menu/create": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "菜单管理"
                ],
                "summary": "创建菜单",
                "parameters": [
                    {
                        "description": "创建菜单",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.CreateMenuRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "desc",
                        "schema": {
                            "$ref": "#/definitions/result.HttpResult"
                        }
                    }
                }
            }
        },
        "/menu/delete/{id}": {
            "delete": {
                "tags": [
                    "菜单管理"
                ],
                "summary": "删除菜单",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "菜单id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "desc",
                        "schema": {
                            "$ref": "#/definitions/result.HttpResult"
                        }
                    }
                }
            }
        },
        "/menu/info/{id}": {
            "get": {
                "tags": [
                    "菜单管理"
                ],
                "summary": "菜单详情",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "菜单id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "desc",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/result.HttpResult"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/models.SysMenu"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
...
        "/menu/update": {
            "patch": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "菜单管理"
                ],
                "summary": "创建菜单",
                "parameters": [
                    {
                        "description": "创建菜单",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.UpdateMenuRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "desc",
                        "schema": {
                            "$ref": "#/definitions/result.HttpResult"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "models.SysDictType": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "description": "创建时间",
                    "type": "string"
                },
                "dictName": {
                    "description": "字典名称",
                    "type": "string"
                },
                "dictType": {
                    "description": "字典类型",
                    "type": "string"
                },
                "id": {
                    "description": "主键",
                    "type": "integer"
                },
                "remark": {
                    "description": "备注",
                    "type": "string"
                },
                "status": {
                    "description": "启用状态 0 启用 1 禁用",
                    "type": "boolean"
                },
                "updatedAt": {
                    "description": "更新时间",
                    "type": "string"
                }
            }
        },
        "models.SysMenu": {
            "type": "object",
            "properties": {
                "component": {
                    "description": "组件路径",
                    "type": "string"
                },
                "createdAt": {
                    "description": "创建时间",
                    "type": "string"
                },
                "id": {
                    "description": "主键",
                    "type": "integer"
                },
                "isCache": {
                    "description": "是否缓存 0缓存 1不缓存",
                    "type": "boolean"
                },
                "isFrame": {
                    "description": "是否外链 0是 1否",
                    "type": "boolean"
                },
                "menuIcon": {
                    "description": "菜单图标",
                    "type": "string"
                },
                "menuName": {
                    "description": "菜单名称",
                    "type": "string"
                },
                "menuSort": {
                    "description": "显示排序",
                    "type": "integer"
                },
                "menuType": {
                    "description": "菜单类型 0目录 1菜单 2按钮",
                    "type": "boolean"
                },
                "parentId": {
                    "description": "父菜单ID",
                    "type": "integer"
                },
                "parentName": {
                    "description": "父菜单名称",
                    "type": "string"
                },
                "perms": {
                    "description": "权限标识",
                    "type": "string"
                },
                "query": {
                    "description": "路由参数",
                    "type": "string"
                },
                "remark": {
                    "description": "备注",
                    "type": "string"
                },
                "router": {
                    "description": "路由地址",
                    "type": "string"
                },
                "status": {
                    "description": "菜单状态 0正常 1停用",
                    "type": "boolean"
                },
                "updatedAt": {
                    "description": "更新时间",
                    "type": "string"
                },
                "visible": {
                    "description": "菜单状态 0显示 1隐藏",
                    "type": "boolean"
                }
            }
        },
...
        "request.CreateDictTypeRequest": {
            "type": "object",
            "properties": {
                "dictName": {
                    "description": "字典名称",
                    "type": "string"
                },
                "dictType": {
                    "description": "字典类型",
                    "type": "string"
                },
                "remark": {
                    "description": "备注",
                    "type": "string"
                },
                "status": {
                    "description": "启用状态 0 启用 1 禁用",
                    "type": "boolean"
                }
            }
        },
        "request.CreateMenuRequest": {
            "type": "object",
            "properties": {
                "component": {
                    "description": "组件路径",
                    "type": "string"
                },
                "isCache": {
                    "description": "是否缓存 0缓存 1不缓存",
                    "type": "boolean"
                },
                "isFrame": {
                    "description": "是否外链 0是 1否",
                    "type": "boolean"
                },
                "menuIcon": {
                    "description": "菜单图标",
                    "type": "string"
                },
                "menuName": {
                    "description": "菜单名称",
                    "type": "string"
                },
                "menuSort": {
                    "description": "显示排序",
                    "type": "integer"
                },
                "menuType": {
                    "description": "菜单类型 0目录 1菜单 2按钮",
                    "type": "boolean"
                },
                "parentId": {
                    "description": "父菜单ID",
                    "type": "integer"
                },
                "parentName": {
                    "description": "父菜单名称",
                    "type": "string"
                },
                "perms": {
                    "description": "权限标识",
                    "type": "string"
                },
                "query": {
                    "description": "路由参数",
                    "type": "string"
                },
                "remark": {
                    "description": "备注",
                    "type": "string"
                },
                "router": {
                    "description": "路由地址",
                    "type": "string"
                },
                "status": {
                    "description": "菜单状态 0正常 1停用",
                    "type": "boolean"
                },
                "visible": {
                    "description": "菜单状态 0显示 1隐藏",
                    "type": "boolean"
                }
            }
        },
...
        "request.UpdateDictTypeRequest": {
            "type": "object",
            "properties": {
                "dictName": {
                    "description": "字典名称",
                    "type": "string"
                },
                "dictType": {
                    "description": "字典类型",
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "remark": {
                    "description": "备注",
                    "type": "string"
                },
                "status": {
                    "description": "启用状态 0 启用 1 禁用",
                    "type": "boolean"
                }
            }
        },
        "request.UpdateMenuRequest": {
            "type": "object",
            "properties": {
                "component": {
                    "description": "组件路径",
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "isCache": {
                    "description": "是否缓存 0缓存 1不缓存",
                    "type": "boolean"
                },
                "isFrame": {
                    "description": "是否外链 0是 1否",
                    "type": "boolean"
                },
                "menuIcon": {
                    "description": "菜单图标",
                    "type": "string"
                },
                "menuName": {
                    "description": "菜单名称",
                    "type": "string"
                },
                "menuSort": {
                    "description": "显示排序",
                    "type": "integer"
                },
                "menuType": {
                    "description": "菜单类型 0目录 1菜单 2按钮",
                    "type": "boolean"
                },
                "parentId": {
                    "description": "父菜单ID",
                    "type": "integer"
                },
                "parentName": {
                    "description": "父菜单名称",
                    "type": "string"
                },
                "perms": {
                    "description": "权限标识",
                    "type": "string"
                },
                "query": {
                    "description": "路由参数",
                    "type": "string"
                },
                "remark": {
                    "description": "备注",
                    "type": "string"
                },
                "router": {
                    "description": "路由地址",
                    "type": "string"
                },
                "status": {
                    "description": "菜单状态 0正常 1停用",
                    "type": "boolean"
                },
                "visible": {
                    "description": "菜单状态 0显示 1隐藏",
                    "type": "boolean"
                }
            }
        },
...
        "response.PageResponse": {
            "type": "object",
            "properties": {
                "list": {},
                "total": {
                    "type": "integer"
                }
            }
        },
...
        "result.HttpResult": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "msg": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}